# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Enterprise Wiki API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tags:
    get:
      operationId: listTags
      tags: [tags]
      summary: List all tags
      parameters:
        - name: q
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of tags with node counts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tag"
    post:
      operationId: createTag
      tags: [tags]
      summary: Create a tag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTag"
      responses:
        "201":
          description: Created tag
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
  /tags/{id}:
    patch:
      operationId: updateTag
      tags: [tags]
      summary: Update a tag
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTag"
      responses:
        "200":
          description: Updated tag
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
    delete:
      operationId: deleteTag
      tags: [tags]
      summary: Delete a tag
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Deleted
  /tags/nodes/{nodeId}:
    get:
      operationId: getNodeTags
      tags: [tags]
      summary: Get tags for a node
      parameters:
        - name: nodeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Node tags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tag"
    post:
      operationId: assignTagToNode
      tags: [tags]
      summary: Assign a tag to a node
      parameters:
        - name: nodeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tagId:
                  type: string
                  format: uuid
              required: [tagId]
      responses:
        "201":
          description: Assigned
  /tags/nodes/{nodeId}/{tagId}:
    delete:
      operationId: removeTagFromNode
      tags: [tags]
      summary: Remove a tag from a node
      parameters:
        - name: nodeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: tagId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Removed
...
components:
...
  schemas:
...
    Tag:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        color:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        nodeCount:
          type: integer
      required:
        - id
        - name
        - slug
    CreateTag:
      type: object
      properties:
        name:
          type: string
        color:
          type: string
          nullable: true
      required:
        - name
...