# lib/api-zod/src/generated/types/createTag.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Enterprise Wiki API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateTag {
  name: string;
  color?: string | null;
}